+++ models/user.py
@@ -5,7 +5,7 @@ from uuid import UUID, uuid4
 from sqlmodel import Field, Relationship, SQLModel
 
 if TYPE_CHECKING:
-    from models.collection import Collection
+    from models.collection import Collection, CollectionPermission
     from models.event import CollectionEvent, DocumentEvent
     from models.update import Update
     from models.paperless import UserPaperless
@@ -55,6 +55,7 @@ class User(UserBase, table=True):
     collections: list["Collection"] = Relationship(back_populates="owner")
     updates: list["Update"] = Relationship(back_populates="user")
     paperless: Optional["UserPaperless"] = Relationship(back_populates="user")
+    permissions: list["CollectionPermission"] = Relationship(back_populates="user")
 
 
 def strip_sensitive(user: UserBase) -> UserSafe:
